daml-lf: move Interface out of reader subpackage
