+++ daml-lf/interface/src/main/scala/com/digitalasset/daml/lf/iface/EnvironmentInterface.scala
@@ -11,16 +11,16 @@ import scala.collection.immutable.Map
 import scalaz.syntax.std.map._
 import scalaz.Semigroup
 
-final case class EnvironmentInterface(typeDecls: Map[Identifier, reader.InterfaceType])
+final case class EnvironmentInterface(typeDecls: Map[Identifier, InterfaceType])
 
 object EnvironmentInterface {
-  def fromReaderInterfaces(i: reader.Interface, o: reader.Interface*): EnvironmentInterface =
+  def fromReaderInterfaces(i: Interface, o: Interface*): EnvironmentInterface =
     EnvironmentInterface((i +: o).flatMap {
-      case reader.Interface(packageId, typeDecls) =>
+      case Interface(packageId, typeDecls) =>
         typeDecls mapKeys (Identifier(packageId, _))
     }(breakOut))
 
-  def fromReaderInterfaces(dar: Dar[reader.Interface]): EnvironmentInterface =
+  def fromReaderInterfaces(dar: Dar[Interface]): EnvironmentInterface =
     fromReaderInterfaces(dar.main, dar.dependencies: _*)
 
   implicit val environmentInterfaceSemigroup: Semigroup[EnvironmentInterface] = Semigroup instance {
