+++ daml-lf/interface/src/main/scala/com/digitalasset/daml/lf/iface/reader/InterfaceReader.scala
@@ -1,7 +1,8 @@
 // Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.daml.lf.iface
+package com.digitalasset.daml.lf
+package iface
 package reader
 
 import ErrorFormatter._
@@ -13,7 +14,6 @@ import scalaz.syntax.apply._
 import scalaz.syntax.bifunctor._
 import scalaz.syntax.monoid._
 import scalaz.syntax.traverse._
-import java.{util => j}
 
 import com.digitalasset.daml.lf.data.ImmArray
 import com.digitalasset.daml.lf.data.ImmArray.ImmArraySeq
@@ -29,54 +29,8 @@ import scala.collection.JavaConverters._
 import scala.collection.breakOut
 import scala.collection.immutable.Map
 
-sealed abstract class InterfaceType extends Product with Serializable {
-  def `type`: DefDataType.FWT
-
-  def fold[Z](normal: DefDataType.FWT => Z, template: (Record.FWT, DefTemplate[Type]) => Z): Z =
-    this match {
-      case InterfaceType.Normal(typ) => normal(typ)
-      case InterfaceType.Template(typ, tpl) => template(typ, tpl)
-    }
-
-  def getType: DefDataType.FWT = `type`
-  def getTemplate: j.Optional[_ <: DefTemplate.FWT] =
-    fold({ _ =>
-      j.Optional.empty()
-    }, { (_, tpl) =>
-      j.Optional.of(tpl)
-    })
-}
-object InterfaceType {
-  final case class Normal(`type`: DefDataType.FWT) extends InterfaceType
-  final case class Template(rec: Record.FWT, template: DefTemplate[Type]) extends InterfaceType {
-    def `type`: DefDataType.FWT = DefDataType(ImmArraySeq.empty, rec)
-  }
-}
-
-/** `templates` stores template info for record types which are templates. Note that all keys of `templates` must be
-  * present in `typeDecls`, too.'
-  */
-final case class Interface(packageId: PackageId, typeDecls: Map[QualifiedName, InterfaceType]) {
-  def getTypeDecls: j.Map[QualifiedName, InterfaceType] = typeDecls.asJava
-}
-
-object Interface {
-  import Errors._, InterfaceReader._
-
-  def read(lf: DamlLf.Archive): (Errors[ErrorLoc, InvalidDataTypeDefinition], Interface) =
-    readInterface(lf)
-
-  def read(lf: (PackageId, DamlLf.ArchivePayload))
-    : (Errors[ErrorLoc, InvalidDataTypeDefinition], Interface) =
-    readInterface(lf)
-
-  def read(f: () => String \/ (PackageId, DamlLf1.Package))
-    : (Errors[ErrorLoc, InvalidDataTypeDefinition], Interface) =
-    readInterface(f)
-}
-
 object InterfaceReader {
-  import Errors._
+  import Errors._, iface.{Interface, InterfaceType}
 
   sealed abstract class InterfaceReaderError extends Product with Serializable
   final case class UnserializableDataType(error: String) extends InterfaceReaderError
